Address Copilot PR review comments
- Fix inconsistent return values in toggleFavorite (add explicit comments)
- Extract swipe gesture magic numbers to named constants (SWIPE_THRESHOLD_X, SWIPE_THRESHOLD_Y)
- Add comprehensive comments explaining Collections tables reserved for future feature
- Clarify intentional non-blocking async calls in admin.js event handlers
- Extract getUserId to shared utility module (user-utils.js) following DRY principle

Resolves all unresolved Copilot review comments in PR #18
